requirements.txt
streamlit>=1.30.0
akshare>=1.12.0
tushare>=1.3.8
pandas>=2.0.0
plotly>=5.18.0
pydantic>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
